+++ projects/packages/my-jetpack/src/class-initializer.php
@@ -31,7 +31,7 @@ class Initializer {
 	 *
 	 * @var string
 	 */
-	const PACKAGE_VERSION = '3.5.0-alpha';
+	const PACKAGE_VERSION = '3.5.0';
 
 	/**
 	 * HTML container ID for the IDC screen on My Jetpack page.