+++ projects/plugins/boost/app/assets/src/js/Main.svelte
@@ -4,6 +4,8 @@
 	import GettingStarted from './pages/getting-started/GettingStarted.svelte';
 	import PurchaseSuccess from './pages/purchase-success/PurchaseSuccess.svelte';
 	import Settings from './pages/settings/Settings.svelte';
+	import config from './stores/config';
+	import { connection } from './stores/connection';
 	import { criticalCssState, isGenerating } from './stores/critical-css-state';
 	import { modulesState } from './stores/modules';
 	import { recordBoostEvent } from './utils/analytics';
@@ -35,10 +37,17 @@
 
 <Router history={routerHistory}>
 	<Route path="upgrade" component={BenefitsInterstitial} />
+	<Route
+		path="getting-started"
+		component={GettingStarted}
+		userConnected={$connection.userConnected}
+		pricing={$config.pricing}
+		isPremium={$config.isPremium}
+		domain={$config.site.domain}
+	/>
 	<Route path="purchase-successful" let:location let:navigate>
 		<PurchaseSuccess {location} {navigate} isImageGuideActive={$modulesState.image_guide.active} />
 	</Route>
-	<Route path="getting-started" component={GettingStarted} />
 	{#if $modulesState.image_size_analysis.available && $modulesState.image_size_analysis.active}
 		<Route path="image-size-analysis/:group/:page" component={RecommendationsPage} />
 	{/if}